must cast long argument to vnode_pager_setsize
SCCS-vsn: sys/ufs/ffs/ffs_vnops.c 7.67

+++ usr/src/sys/ufs/ffs/ffs_vnops.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ffs_vnops.c	7.66 (Berkeley) %G%
+ *	@(#)ffs_vnops.c	7.67 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -298,7 +298,7 @@ ffs_write(vp, uio, ioflag, cred)
 		bn = bp->b_blkno;
 		if (uio->uio_offset + n > ip->i_size) {
 			ip->i_size = uio->uio_offset + n;
-			vnode_pager_setsize(vp, ip->i_size);
+			vnode_pager_setsize(vp, (u_long)ip->i_size);
 		}
 		size = blksize(fs, ip, lbn);
 		(void) vnode_pager_uncache(vp);
